* commit '91cf14baade6866910ce71a7e935a0a42a24b2a7':
  Small cleanup due to fortran module name changes
  Add default arguments to the yaml declarations
  Fixing some small style things and removing some overloaded methods in by adding default arguments
  Add documentation comments
  Add Fortran interface and tests for preserving contents during a load from file.
  Add preserve_contents option in sidre and spio to allow for loading into a Group without clearing the existing contents of the Group.

+++ src/components/sidre/src/core/Group.cpp
@@ -1374,7 +1374,8 @@ void Group::save(const hid_t& h5_id,
  *************************************************************************
  */
 void Group::load(const std::string& path,
-                     const std::string& protocol)
+                 const std::string& protocol,
+                 bool preserve_contents)
 {
   std::string new_name; 
   if (protocol == "sidre_hdf5")
@@ -1382,7 +1383,7 @@ void Group::load(const std::string& path,
     Node n;
     conduit::relay::io::load(path,"hdf5", n);
     SLIC_ASSERT(n.has_path("sidre"));
-    importFrom(n["sidre"]);
+    importFrom(n["sidre"], preserve_contents);
     if (n.has_path("sidre_group_name")) {
       new_name = n["sidre_group_name"].as_string();
     }
@@ -1392,7 +1393,7 @@ void Group::load(const std::string& path,
     Node n;
     conduit::relay::io::load(path,"conduit_json", n);
     SLIC_ASSERT(n.has_path("sidre"));
-    importFrom(n["sidre"]);
+    importFrom(n["sidre"], preserve_contents);
     if (n.has_path("sidre_group_name")) {
       new_name = n["sidre_group_name"].as_string();
     }
@@ -1402,7 +1403,7 @@ void Group::load(const std::string& path,
     Node n;
     conduit::relay::io::load(path,"json", n);
     SLIC_ASSERT(n.has_path("sidre"));
-    importFrom(n["sidre"]);
+    importFrom(n["sidre"], preserve_contents);
     if (n.has_path("sidre_group_name")) {
       new_name = n["sidre_group_name"].as_string();
     }
@@ -1411,7 +1412,7 @@ void Group::load(const std::string& path,
   {
     Node n;
     conduit::relay::io::load(path,"hdf5", n);
-    importConduitTree(n);
+    importConduitTree(n, preserve_contents);
     if (n.has_path("sidre_group_name")) {
       new_name = n["sidre_group_name"].as_string();
     }
@@ -1422,7 +1423,7 @@ void Group::load(const std::string& path,
   {
     Node n;
     conduit::relay::io::load(path,protocol, n);
-    importConduitTree(n);
+    importConduitTree(n, preserve_contents);
     if (n.has_path("sidre_group_name")) {
       new_name = n["sidre_group_name"].as_string();
     }
@@ -1443,7 +1444,8 @@ void Group::load(const std::string& path,
  *************************************************************************
  */
 void Group::load(const hid_t& h5_id,
-                     const std::string &protocol)
+                 const std::string &protocol,
+                 bool preserve_contents)
 {
   // supported here:
   // "sidre_hdf5"
@@ -1454,7 +1456,7 @@ void Group::load(const hid_t& h5_id,
     Node n;
     conduit::relay::io::hdf5_read(h5_id,n);
     SLIC_ASSERT(n.has_path("sidre"));
-    importFrom(n["sidre"]);
+    importFrom(n["sidre"], preserve_contents);
     if (n.has_path("sidre_group_name")) {
       new_name = n["sidre_group_name"].as_string();
     }
@@ -1464,7 +1466,7 @@ void Group::load(const hid_t& h5_id,
     SLIC_ERROR("Protocol " << protocol << " not yet supported for file load.");
     Node n;
     conduit::relay::io::hdf5_read(h5_id, n);
-    importConduitTree(n);
+    importConduitTree(n, preserve_contents);
     if (n.has_path("sidre_group_name")) {
       new_name = n["sidre_group_name"].as_string();
     }
@@ -1843,14 +1845,17 @@ void Group::exportTo(conduit::Node& result,
  *************************************************************************
  */
 
-void Group::importFrom(conduit::Node & node)
+void Group::importFrom(conduit::Node & node, bool preserve_contents)
 {
   // TODO - May want to put in a little meta-data into these files like a 'version'
   // or tag identifying the data.  We don't want someone giving us a file that
   // doesn't have our full multiView->buffer connectivity in there.
 
-  destroyGroups();
-  destroyViews();
+  if (!preserve_contents)
+  {
+    destroyGroups();
+    destroyViews();
+  }
 
   // First - Import Buffers into the DataStore.
   std::map<IndexType, IndexType> buffer_indices_map;
@@ -1894,7 +1899,7 @@ void Group::importFrom(conduit::Node & node)
  *************************************************************************
  */
 void Group::importFrom(conduit::Node& node,
-                           const std::map<IndexType, IndexType>& buffer_id_map)
+                       const std::map<IndexType, IndexType>& buffer_id_map)
 {
   if ( node.has_path("views") )
   {
@@ -1934,10 +1939,13 @@ void Group::importFrom(conduit::Node& node,
  *************************************************************************
  */
 
-void Group::importConduitTree(conduit::Node &node)
+void Group::importConduitTree(conduit::Node &node, bool preserve_contents)
 {
-  destroyGroups();
-  destroyViews();
+  if (!preserve_contents)
+  {
+    destroyGroups();
+    destroyViews();
+  }
 
   //
   DataType node_dtype = node.dtype();
@@ -1954,7 +1962,7 @@ void Group::importConduitTree(conduit::Node &node)
       {
         // create group
         Group * grp = createGroup(cld_name);
-        grp->importConduitTree(cld_node);
+        grp->importConduitTree(cld_node, preserve_contents);
       }
       else if(cld_dtype.is_empty())
       {
